PosNegGTSR/requirements.txt
addict
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tensorboard
tqdm
yapf
torchprofile
pytorch-fid
lpips
matplotlib
jupyter
wandb
timm
sk-video>=1.1.10
moviepy>=1.0.3
einops
dists-pytorch
